requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
flasgger==0.9.7.1
psycopg2-binary==2.9.7
pymongo==4.5.0
firebase-admin==6.4.0
